requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
google-generativeai
langchain
langchain-google-genai
python-multipart
httpx
requests